Fix regression in d82533a that broke offline support
The original fix assumes that the app is served from the root of the
domain. However, the app is served from a subdirectory on the domain,
breaking the templating because the precached route is never matched.
Using '.' to refer to the current subdirectory (rather than the root of
the domain, '/') fixes the issue in all cases.