Remove packageManager declaration
Since the caret (for semver) seems to be unsupported in Node v16, I
would prefer to drop this field, especially since corepack is still
experimental. Plus, keeping it would mean that I could not opt into
using the latest version of pnpm automatically (i.e. the version would
be fixed).